Helper-files
Created a helper-files folder inside the app folder of the Angular project and added a "content-interface.ts" file to define an interface called "Content".
Added "content-list.ts" file to define a class called "ContentList" which holds an array of type Content, with methods such as getter function that returns the Content array, add function that adds Content item to the end of the array, a function that returns the number of items in the array and a function that takes an input of an index of the array and returns a reader-friendly html output of a Content item's properties at that index.