Angular assignment 2
Created a new component called Content-list. In this component, defined an array of 6 Content items.
In the Content-list component's HTML file, used an ngFor directive to display 5 Content-card components. Each Content-card component received the values from the Content array as input. Details displayed in each Content-card component:
-The first Content-card component displayed have a 2-pixel solid black border, all the other 1px
-Title displayed in a header tag.
-Description, creator, and image displayed inside the card.
The positioning and styling of these elements changed.
Clicking on the image display the Content's id and title in the console.

BONUS: Added an If statement: when the values of type, image, or tags are not set,they are not displayed on the Content-card component.